backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
pydantic==2.6.1
python-multipart==0.0.9
httpx==0.26.0
pyyaml==6.0.1
aiofiles==23.2.1
requests==2.31.0
pytest==7.4.4
pytest-asyncio==0.23.4
